- Now allow Monotone style URLs when connecting to remote servers. - Updated the copyright information in the about dialog. - Updated the MAS dependency in the installer. - The history graph button is now enabled for remote connections. - Updated the README file with new dependency information and an explanation of   the `Use Gtk2::Labels' installer option. - Updated the help accordingly, also making some unrelated corrections (mainly   missing links). - Updated the version number to 1.0.

+++ lib/perl/ManageServerBookmarks.pm
@@ -204,15 +204,17 @@ sub add_server_button_clicked_cb($$)
     # Check entry to see if it is valid.
 
     $server = $instance->{server_entry}->get_text();
-    if ($server !~ m/^[A-Za-z0-9._-]+(:\d+)?$/)
+    if ($server !~ m/^[A-Za-z0-9._-]+(:\d+)?$/
+        && $server !~ m/\//)
     {
         my $dialog = Gtk2::MessageDialog->new
             ($instance->{window},
              ["modal"],
              "warning",
              "close",
-             __x("`{server}' is an invalid server\n"
-                 . "name (<server>[:port] is expected).",
+             __x("`{server}' is invalid. Expected either\n"
+                     . "<Server>[:<Port>] or a Monotone URL such as\n"
+                     . "`mtn://code.monotone.ca/mtn-browse'.",
                  server => $server));
         busy_dialog_run($dialog);
         $dialog->destroy();
